P1: DFS: Add DFS implementation
The code is for the most part copied over from BFS
Only differences is that we use a stack instead of queue,
and the order of visiting nodes is reversed (since it's a
stack).

This patch also fixes certain nits that were causing off-by-one
errors with the expected ouput vis a vis what is expected by
the scripts.